Move curses app into toot.ui module
